+++ examples/stream_chat_with_tools.rs
@@ -153,6 +153,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
             StreamEvent::Error(e) => {
                 eprintln!("\nStream error: {}", e);
             }
+            _ => {}
         }
     }